feat(deps): RustCrypto 0.10→0.11 + aws-lc-rs removal (Plan 126 B4/C1) (#1686)
* feat(deps): RustCrypto 0.10→0.11 migration + aws-lc-rs removal (Plan 126 B4/C1)

Migrates the workspace RustCrypto stack from the 0.10 generation to 0.11
and uses the resulting compatibility to remove aws-lc-rs from the
dependency tree entirely.

RustCrypto 0.11 migration:
- sha2 0.10→0.11, digest 0.10→0.11, hmac 0.12→0.13, hkdf 0.12→0.13
- aes-gcm 0.10→0.11, ed25519-dalek 2.2→3.0 (aead/cipher generation bump)
- sha2 0.11 drops std::io::Write; replace io::copy-based callers with
  manual read loops + hex::encode(hasher.finalize()) at every site
- hmac 0.13 moves new_from_slice to KeyInit trait; add KeyInit to imports
- ed25519-dalek 3.0 uses rand_core 0.10; avoid the version conflict by
  using rand::RngCore::fill_bytes + SigningKey::from_bytes everywhere
  instead of SigningKey::generate (pattern applied across ~40 call sites)
- aes-gcm 0.11 removes OsRng/AeadCore re-exports; generate nonce via
  rand::RngCore::fill_bytes + Nonce::from([u8; 12])

aws-lc-rs removal:
- Bump workspace reqwest 0.12→0.13 with rustls-no-provider feature;
  no longer pulls rustls-platform-verifier (the aws-lc-rs source)
- Install ring TLS provider at startup in mvm-cli/src/commands/mod.rs
- cargo tree -i aws-lc-rs returns empty; cmake build gone from the tree

Dependency ratchet:
- deny.toml: document pre-existing rand/rand_core dual-major as skip
  entries (hickory-proto 0.26 already pulled rand 0.10 before this
  change; we net-reduce from 3 versions to 2)
- All workspace gates pass: cargo nextest (1317), clippy -D warnings,
  fmt --all --check, cargo deny check, xtask check-no-spec-refs-in-comments

Plan 126 B4/C1 acceptance criteria met; plan + refactor-status updated.

* fix(rustcrypto): remove stale OsRng imports

* fix(client): install rustls provider in gateway tests

* fix(template): install rustls provider for local probe

* fix(http): install rustls provider for hardened clients

* Install rustls provider in loopback test

* Install rustls provider in web fetch client

* Install rustls provider in registry client

---------

+++ crates/mvm-backend/src/vmm/virtio_fs.rs
@@ -438,7 +438,7 @@ fn encode_attr(ino: u64, path: &Path) -> Option<Vec<u8>> {
     } else {
         S_IFREG
     };
-    let perm = (md.mode() & 0o7777) as u32;
+    let perm = md.mode() & 0o7777;
     let mut b = Vec::new();
     put_u64(&mut b, ino); // ino
     put_u64(&mut b, md.len()); // size
